adding recursive clustering_algorithm
